Refactor supplementaryQueries into a mixin
