+++ src/datasource/supplementaryQueries.ts
@@ -1,0 +1,290 @@
+import {
+  DataFrame,
+  DataQueryRequest,
+  DataQueryResponse,
+  DataSourceApi,
+  DataSourceJsonData,
+  DataSourceWithSupplementaryQueriesSupport,
+  FieldColorModeId,
+  FieldType,
+  LoadingState,
+  LogLevel,
+  LogsVolumeCustomMetaData,
+  LogsVolumeType,
+  SupplementaryQueryType,
+} from '@grafana/data';
+import { BarAlignment, DataQuery, GraphDrawStyle, StackingMode } from "@grafana/schema";
+import { colors } from "@grafana/ui";
+import { getIntervalInfo } from '@/utils/time';
+import { cloneDeep, groupBy } from "lodash";
+import { Observable, isObservable, from } from 'rxjs';
+import { BucketAggregation, ElasticsearchQuery } from '@/types';
+import { BaseQuickwitDataSourceConstructor } from './base';
+
+export const REF_ID_STARTER_LOG_VOLUME = 'log-volume-';
+
+export function withSupplementaryQueries<T extends BaseQuickwitDataSourceConstructor> ( Base: T ){
+  return class DSWithSupplementaryQueries extends Base implements DataSourceWithSupplementaryQueriesSupport<ElasticsearchQuery> {
+  /**
+   * Returns an observable that will be used to fetch supplementary data based on the provided
+   * supplementary query type and original request.
+   */
+  getDataProvider(
+    type: SupplementaryQueryType,
+    request: DataQueryRequest<ElasticsearchQuery>
+  ): Observable<DataQueryResponse> | undefined {
+    if (!this.getSupportedSupplementaryQueryTypes().includes(type)) {
+      return undefined;
+    }
+    switch (type) {
+      case SupplementaryQueryType.LogsVolume:
+        return this.getLogsVolumeDataProvider(request);
+      default:
+        return undefined;
+    }
+  }
+
+  /**
+   * Returns supplementary query types that data source supports.
+   */
+  getSupportedSupplementaryQueryTypes(): SupplementaryQueryType[] {
+    return [SupplementaryQueryType.LogsVolume];
+  }
+
+  /**
+   * Returns a supplementary query to be used to fetch supplementary data based on the provided type and original query.
+   * If provided query is not suitable for provided supplementary query type, undefined should be returned.
+   */
+  // FIXME: options should be of type SupplementaryQueryOptions but this type is not public.
+  getSupplementaryQuery(options: any, query: ElasticsearchQuery): ElasticsearchQuery | undefined {
+    if (!this.getSupportedSupplementaryQueryTypes().includes(options.type)) {
+      return undefined;
+    }
+
+    let isQuerySuitable = false;
+
+    switch (options.type) {
+      case SupplementaryQueryType.LogsVolume:
+        // it has to be a logs-producing range-query
+        isQuerySuitable = !!(query.metrics?.length === 1 && query.metrics[0].type === 'logs');
+        if (!isQuerySuitable) {
+          return undefined;
+        }
+        const bucketAggs: BucketAggregation[] = [];
+        const timeField = this.timeField ?? 'timestamp';
+
+        if (this.logLevelField) {
+          bucketAggs.push({
+            id: '2',
+            type: 'terms',
+            settings: {
+              min_doc_count: '0',
+              size: '0',
+              order: 'desc',
+              orderBy: '_count',
+            },
+            field: this.logLevelField,
+          });
+        }
+        bucketAggs.push({
+          id: '3',
+          type: 'date_histogram',
+          settings: {
+            interval: 'auto',
+            min_doc_count: '0',
+            trimEdges: '0',
+          },
+          field: timeField,
+        });
+
+        return {
+          refId: `${REF_ID_STARTER_LOG_VOLUME}${query.refId}`,
+          query: query.query,
+          metrics: [{ type: 'count', id: '1' }],
+          bucketAggs,
+        };
+
+      default:
+        return undefined;
+    }
+  }
+
+  getLogsVolumeDataProvider(request: DataQueryRequest<ElasticsearchQuery>): Observable<DataQueryResponse> | undefined {
+    const logsVolumeRequest = cloneDeep(request);
+    const targets = logsVolumeRequest.targets
+      .map((target) => this.getSupplementaryQuery({ type: SupplementaryQueryType.LogsVolume }, target))
+      .filter((query): query is ElasticsearchQuery => !!query);
+
+    if (!targets.length) {
+      return undefined;
+    }
+
+    return queryLogsVolume(
+      this,
+      { ...logsVolumeRequest, targets },
+      {
+        range: request.range,
+        targets: request.targets,
+        extractLevel: (dataFrame: any) => getLogLevelFromKey(dataFrame || ''),
+      }
+    );
+  }
+  };
+}
+
+// Copy/pasted from grafana/data as it is deprecated there.
+function getLogLevelFromKey(dataframe: DataFrame): LogLevel {
+  const name = dataframe.fields[1].config.displayNameFromDS || ``;
+  const level = (LogLevel as any)[name.toString().toLowerCase()];
+  if (level) {
+    return level;
+  }
+  return LogLevel.unknown;
+}
+
+/**
+ * Creates an observable, which makes requests to get logs volume and aggregates results.
+ */
+
+export function queryLogsVolume<TQuery extends DataQuery, TOptions extends DataSourceJsonData>(
+  datasource: DataSourceApi<TQuery, TOptions>,
+  logsVolumeRequest: DataQueryRequest<TQuery>,
+  options: any
+): Observable<DataQueryResponse> {
+  const timespan = options.range.to.valueOf() - options.range.from.valueOf();
+  const intervalInfo = getIntervalInfo(timespan, 400);
+
+  logsVolumeRequest.interval = intervalInfo.interval;
+  logsVolumeRequest.scopedVars.__interval = { value: intervalInfo.interval, text: intervalInfo.interval };
+
+  if (intervalInfo.intervalMs !== undefined) {
+    logsVolumeRequest.intervalMs = intervalInfo.intervalMs;
+    logsVolumeRequest.scopedVars.__interval_ms = { value: intervalInfo.intervalMs, text: intervalInfo.intervalMs };
+  }
+
+  logsVolumeRequest.hideFromInspector = true;
+
+  return new Observable((observer) => {
+    let logsVolumeData: DataFrame[] = [];
+    observer.next({
+      state: LoadingState.Loading,
+      error: undefined,
+      data: [],
+    });
+
+    const queryResponse = datasource.query(logsVolumeRequest);
+    const queryObservable = isObservable(queryResponse) ? queryResponse : from(queryResponse);
+
+    const subscription = queryObservable.subscribe({
+      complete: () => {
+        observer.complete();
+      },
+      next: (dataQueryResponse: DataQueryResponse) => {
+        const { error } = dataQueryResponse;
+        if (error !== undefined) {
+          observer.next({
+            state: LoadingState.Error,
+            error,
+            data: [],
+          });
+          observer.error(error);
+        } else {
+          const framesByRefId = groupBy(dataQueryResponse.data, 'refId');
+          logsVolumeData = dataQueryResponse.data.map((dataFrame) => {
+            let sourceRefId = dataFrame.refId || '';
+            if (sourceRefId.startsWith('log-volume-')) {
+              sourceRefId = sourceRefId.substr('log-volume-'.length);
+            }
+
+            const logsVolumeCustomMetaData: LogsVolumeCustomMetaData = {
+              logsVolumeType: LogsVolumeType.FullRange,
+              absoluteRange: { from: options.range.from.valueOf(), to: options.range.to.valueOf() },
+              datasourceName: datasource.name,
+              sourceQuery: options.targets.find((dataQuery: any) => dataQuery.refId === sourceRefId)!,
+            };
+
+            dataFrame.meta = {
+              ...dataFrame.meta,
+              custom: {
+                ...dataFrame.meta?.custom,
+                ...logsVolumeCustomMetaData,
+              },
+            };
+            return updateLogsVolumeConfig(dataFrame, options.extractLevel, framesByRefId[dataFrame.refId].length === 1);
+          });
+
+          observer.next({
+            state: dataQueryResponse.state,
+            error: undefined,
+            data: logsVolumeData,
+          });
+        }
+      },
+      error: (error: any) => {
+        observer.next({
+          state: LoadingState.Error,
+          error: error,
+          data: [],
+        });
+        observer.error(error);
+      },
+    });
+    return () => {
+      subscription?.unsubscribe();
+    };
+  });
+}
+const updateLogsVolumeConfig = (
+  dataFrame: DataFrame,
+  extractLevel: (dataFrame: DataFrame) => LogLevel,
+  oneLevelDetected: boolean
+): DataFrame => {
+  dataFrame.fields = dataFrame.fields.map((field) => {
+    if (field.type === FieldType.number) {
+      field.config = {
+        ...field.config,
+        ...getLogVolumeFieldConfig(extractLevel(dataFrame), oneLevelDetected),
+      };
+    }
+    return field;
+  });
+  return dataFrame;
+};
+const LogLevelColor = {
+  [LogLevel.critical]: colors[7],
+  [LogLevel.warning]: colors[1],
+  [LogLevel.error]: colors[4],
+  [LogLevel.info]: colors[0],
+  [LogLevel.debug]: colors[5],
+  [LogLevel.trace]: colors[2],
+  [LogLevel.unknown]: '#8e8e8e' // or '#bdc4cd',
+};
+/**
+ * Returns field configuration used to render logs volume bars
+ */
+function getLogVolumeFieldConfig(level: LogLevel, oneLevelDetected: boolean) {
+  const name = oneLevelDetected && level === LogLevel.unknown ? 'logs' : level;
+  const color = LogLevelColor[level];
+  return {
+    displayNameFromDS: name,
+    color: {
+      mode: FieldColorModeId.Fixed,
+      fixedColor: color,
+    },
+    custom: {
+      drawStyle: GraphDrawStyle.Bars,
+      barAlignment: BarAlignment.Center,
+      lineColor: color,
+      pointColor: color,
+      fillColor: color,
+      lineWidth: 1,
+      fillOpacity: 100,
+      stacking: {
+        mode: StackingMode.Normal,
+        group: 'A',
+      },
+    },
+  };
+}
+
+
