+++ src/utils/mixins.ts
@@ -1,0 +1,1 @@
+export type GConstructor<T = {}> = new (...args: any[]) => T;
